Refactor PLong component with useReducer and update use
In the PLong component, useState was replaced with useReducer to simplify state management and make the component easier to maintain. Two actions, 'TICK' and 'RESET', were added to the reducer function. PLong's visibility behavior was also updated in the Home screen, making the extended biographies visible only when 'isAboutExtended' is true.